refactor from user to auth
