+++ server/routes/auth.js
@@ -1,7 +1,7 @@
 const express = require('express');
 const router = express.Router();
 
-const { signup, signin, signout, requireSignin } = require('../controllers/user');
+const { signup, signin, signout, requireSignin } = require('../controllers/auth');
 const { userSignupValidator } = require('../validator');
 
 router.post('/signup', userSignupValidator, signup);
